Work on RV32 (very WIP)

+++ src/device/clint.rs
@@ -3,9 +3,9 @@ use crate::hw::param;
 use crate::hw::riscv;
 
 /// Get the current CLINT time.
-pub fn read_mtime() -> u64 {
-    let base = param::CLINT_BASE as *mut u64;
-    let mtime: u64;
+pub fn read_mtime() -> usize {
+    let base = param::CLINT_BASE as *mut usize;
+    let mtime: usize;
     unsafe {
         mtime = base.byte_add(0xBFF8).read_volatile();
     }
@@ -17,8 +17,8 @@ pub fn read_mtime() -> u64 {
 /// a *machine*-mode interrupt.
 // mtimecmp reg is at base + 0x4000
 // mtime reg is base + 0xbff8
-pub fn set_mtimecmp(interval: u64) {
-    let hartid = riscv::get_mhartid() as usize;
+pub fn set_mtimecmp(interval: usize) {
+    let hartid = riscv::get_mhartid();
     let base = param::CLINT_BASE as *mut usize;
     unsafe {
         // One mtime register for all cores.
